requirements.txt
alembic==1.13.2                                                                                                                                                                                                                               
aniso8601==9.0.1                                                                                                                                                                                                                              
arrow==1.3.0                                                                                                                                                                                                                                  
attrs==23.2.0                                                                                                                                                                                                                                 
autoenv==1.0.0                                                                                                                                                                                                                                
blinker==1.8.2                                                                                                                                                                                                                                
certifi==2024.7.4                                                                                                                                                                                                                             
cffi==1.16.0                                               
chardet==5.2.0                                             
charset-normalizer==3.3.2                                  
click==8.1.7                                               
cryptography==42.0.8                                       
dnspython==2.6.1                                           
dominate==2.9.1                                            
Flask==3.0.3                                               
Flask-Bootstrap==3.3.7.1                                   
Flask-Cors==4.0.1                                          
Flask-DotEnv==0.1.2                                        
Flask-Login==0.6.3                                         
Flask-Migrate==4.0.7                                       
flask-restx==1.3.0                                         
Flask-SQLAlchemy==3.1.1                                    
Flask-Uploads==0.2.1                                       
Flask-WTF==1.2.1                                           
greenlet==3.0.3                                            
gunicorn==22.0.0                                           
idna==3.7                                                  
importlib_resources==6.4.0                                 
itsdangerous==2.2.0                                        
Jinja2==3.1.4                                              
jsonschema==4.22.0                                         
jsonschema-specifications==2023.12.1                       
Mako==1.3.5                                                
MarkupSafe==2.1.5                                          
marshmallow==3.21.3                                        
packaging==24.1                                            
passlib==1.7.4                                             
protobuf==5.27.2
pycparser==2.22
pycryptodome==3.20.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
SQLAlchemy==2.0.31
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
numpy==2.0.0
pandas==2.2.2
shapely==2.0.4
fiona==1.9.6
six==1.16.0
cython==3.0.10
pyproj==3.6.1
geopandas==1.0.1